requirements.txt
superduperreload-core==0.0.2
ipykernel
ipyflow-core
ipython
